Handle Chrome double escape behaviour
resolves #4365

Pressing escape key twice will close open dialogs in Chrome no matter
how onCancel is configured https://issues.chromium.org/issues/346597066

To handle that we move the onClickBackButton function one component
higher to add it as onClose prop to the Dialog and handle the event
the same way a BackButton Click would be handled
