+++ packages/frontend/src/components/screens/WelcomeScreen/index.tsx
@@ -6,6 +6,12 @@ import InstantOnboardingScreen from './InstantOnboardingScreen'
 import OnboardingScreen from './OnboardingScreen'
 import useInstantOnboarding from '../../../hooks/useInstantOnboarding'
 import { getConfiguredAccounts } from '../../../backend/account'
+import { BackendRemote, EffectfulBackendActions } from '../../../backend-com'
+import useDialog from '../../../hooks/dialog/useDialog'
+import AlertDialog from '../../dialogs/AlertDialog'
+import { getLogger } from '@deltachat-desktop/shared/logger'
+
+const log = getLogger('renderer/components/screens/WelcomScreen')
 
 type Props = {
   selectedAccountId: number
@@ -20,6 +26,7 @@ export default function WelcomeScreen({ selectedAccountId, ...props }: Props) {
     startInstantOnboardingFlow,
   } = useInstantOnboarding()
   const [hasConfiguredAccounts, setHasConfiguredAccounts] = useState(false)
+  const { openDialog } = useDialog()
   const showBackButton = hasConfiguredAccounts
 
   useLayoutEffect(() => {
@@ -35,6 +42,36 @@ export default function WelcomeScreen({ selectedAccountId, ...props }: Props) {
     checkAccounts()
   }, [])
 
+  /**
+   * this function is called when the back button is clicked
+   * but also if the dialog is closed by pressing esc multiple
+   * times, which will force Chrome to close the dialog
+   * see https://issues.chromium.org/issues/346597066
+   *
+   * it will cancel the account creation process and call
+   * onExitWelcomeScreen
+   */
+  const onClickBackButton = async () => {
+    try {
+      const acInfo = await BackendRemote.rpc.getAccountInfo(selectedAccountId)
+      if (acInfo.kind === 'Unconfigured') {
+        await props.onUnSelectAccount()
+        await EffectfulBackendActions.removeAccount(selectedAccountId)
+      }
+      props.onExitWelcomeScreen()
+    } catch (error) {
+      if (error instanceof Error) {
+        openDialog(AlertDialog, {
+          message: error?.message,
+          cb: () => {},
+        })
+      } else {
+        log.error('unexpected error type', error)
+        throw error
+      }
+    }
+  }
+
   return (
     <ImageBackdrop variant='welcome'>
       <Dialog
@@ -42,6 +79,7 @@ export default function WelcomeScreen({ selectedAccountId, ...props }: Props) {
         width={400}
         canEscapeKeyClose={false}
         canOutsideClickClose={false}
+        onClose={onClickBackButton}
       >
         {!showInstantOnboarding ? (
           <OnboardingScreen
@@ -49,6 +87,7 @@ export default function WelcomeScreen({ selectedAccountId, ...props }: Props) {
             selectedAccountId={selectedAccountId}
             showBackButton={showBackButton}
             hasConfiguredAccounts={hasConfiguredAccounts}
+            onClickBackButton={onClickBackButton}
             {...props}
           />
         ) : (
